Refactor - introduce NomisClient::Base
